+ toggleable visual Debug Info + Removed obsolete FuelSystem class
