+++ Cockpit/Scripts/Systems/Avionics.lua
@@ -5,63 +5,56 @@ local sensor_data = get_base_data()
 local update_time_step = 0.05  
 make_default_activity(update_time_step)
 
-local meters_to_feet = 3.2808399
+local meters_to_feet            = 3.2808399
 local feet_per_meter_per_minute = meters_to_feet * 60
-local radian_to_degree = 57.2957795
-local KG_TO_POUNDS = 2.20462
+local radian_to_degree          = 57.2957795
+local KG_TO_POUNDS              = 2.20462
+local lowalt                    = -0.1
 
 
 local ALT_PRESSURE_MAX = 30.99 -- in Hg
 local ALT_PRESSURE_MIN = 29.10 -- in Hg
 local ALT_PRESSURE_STD = 29.92 -- in Hg
+local alt_setting        = ALT_PRESSURE_STD
 
-local alt_10k = get_param_handle("ALT_10000") -- 0 to 100,000
-local alt_1k = get_param_handle("ALT_1000") -- 0 to 10,000
-local alt_100s = get_param_handle("ALT_100") -- 0 to 1000
-local alt_adj_Nxxx = get_param_handle("ALT_ADJ_Nxxx") -- 1st digit, 0-10 is input
-local alt_adj_xNxx = get_param_handle("ALT_ADJ_xNxx") -- 2nd digit, 0-10 input
-local alt_adj_xxNx = get_param_handle("ALT_ADJ_xxNx") -- 3rd digit, 0-10 input
-local alt_adj_xxxN = get_param_handle("ALT_ADJ_xxxN") -- 4th digit, 0-10 input
-local alt_adj_MBNxxx = get_param_handle("ALT_ADJ_MBNxxx") -- 
-local alt_adj_MBxNxx = get_param_handle("ALT_ADJ_MBxNxx") -- 
-local alt_adj_MBxxNx = get_param_handle("ALT_ADJ_MBxxNx") -- 
-local alt_adj_MBxxxN = get_param_handle("ALT_ADJ_MBxxxN") -- 
-local current_Ralt = get_param_handle("CURRENT_RALT")
-local Ralt_Off = get_param_handle("RALT_OFF")
-local current_hdg = get_param_handle("CURRENT_HDG")
-local current_fuelT  = get_param_handle("CURRENT_FUELT")
+local alt_10k            = get_param_handle("ALT_10000") -- 0 to 100,     000
+local alt_1k             = get_param_handle("ALT_1000") -- 0 to 10,       000
+local alt_100s           = get_param_handle("ALT_100") -- 0 to 1000
+local alt_adj_Nxxx       = get_param_handle("ALT_ADJ_Nxxx") -- 1st digit, 0-10 is input
+local alt_adj_xNxx       = get_param_handle("ALT_ADJ_xNxx") -- 2nd digit, 0-10 input
+local alt_adj_xxNx       = get_param_handle("ALT_ADJ_xxNx") -- 3rd digit, 0-10 input
+local alt_adj_xxxN       = get_param_handle("ALT_ADJ_xxxN") -- 4th digit, 0-10 input
+local alt_adj_MBNxxx     = get_param_handle("ALT_ADJ_MBNxxx") --
+local alt_adj_MBxNxx     = get_param_handle("ALT_ADJ_MBxNxx") --
+local alt_adj_MBxxNx     = get_param_handle("ALT_ADJ_MBxxNx") --
+local alt_adj_MBxxxN     = get_param_handle("ALT_ADJ_MBxxxN") --
+local current_Ralt       = get_param_handle("CURRENT_RALT")
+local Ralt_Off           = get_param_handle("RALT_OFF")
+local current_hdg        = get_param_handle("CURRENT_HDG")
+local current_fuelT      = get_param_handle("CURRENT_FUELT")
+local current_tas        = get_param_handle("CURRENT_TAS")
+local current_asi        = get_param_handle("CURRENT_ASI")
+local current_pitch      = get_param_handle("CURRENT_PITCH")
+local current_roll       = get_param_handle("CURRENT_ROLL")
+local test               = get_param_handle("TEST_PARAM")
+local ROLL_F             = get_param_handle("ROLL_F")
+local YAM_F              = get_param_handle("YAM_F")
+local HMD_STATE          = get_param_handle("HMD_STATE")
+local LOWALT             = get_param_handle("LOWALT")
+local DEBUG_INFO_ENABLED = get_param_handle("DEBUG_INFO_ENABLED")
 
-local current_tas = get_param_handle("CURRENT_TAS")
-local current_asi = get_param_handle("CURRENT_ASI")
-
-local current_pitch = get_param_handle("CURRENT_PITCH")
-local current_roll = get_param_handle("CURRENT_ROLL")
-
-local test = get_param_handle("TEST_PARAM")
-
-local ROLL_F = get_param_handle("ROLL_F")
-local YAM_F = get_param_handle("YAM_F")
-
-
-local alt_setting = ALT_PRESSURE_STD
-local HMD_STATE = get_param_handle("HMD_STATE")
-
-local lowalt = -0.1
-local LOWALT = get_param_handle("LOWALT")
-
-
-
-alt_10k:set(0.0)
-alt_1k:set(0.0)
-alt_100s:set(0.0)
-current_Ralt:set(0)
-Ralt_Off:set(0)
-current_hdg:set(0)
-current_fuelT:set(0)
-
-HMD_STATE:set(0)
 
 function post_initialize()
+    alt_10k:set(0.0)
+    alt_1k:set(0.0)
+    alt_100s:set(0.0)
+    current_Ralt:set(0)
+    Ralt_Off:set(0)
+    current_hdg:set(0)
+    current_fuelT:set(0)
+    HMD_STATE:set(0)
+    DEBUG_INFO_ENABLED:set(1)
+
 	current_fuelT:set(sensor_data.getTotalFuelWeight()*KG_TO_POUNDS)
 	current_hdg:set(360-(sensor_data.getHeading()*radian_to_degree))
 	update_altimeter()
@@ -79,6 +72,7 @@ dev:listen_command(device_commands.AltimeterSet)
 dev:listen_command(3571)
 dev:listen_command(3581)
 dev:listen_command(1507)
+dev:listen_command(EFM_commands.ToggleDebugInfo)
 -- dev:listen_command(2143)
 
 -- dev:listen_command(49) -- vertical
@@ -91,53 +85,6 @@ vs= 350
 angle = 60
 h0 = 20
 g=9.80665
-
-
-function SetCommand(command,value)  
--- print_message_to_user("command,value="..command) 
-   if command == 1501 then
-	h  = h0 + ( math.pow(vs * math.sin(math.rad(angle)), 2) / (2 * g))
-	ve = math.sqrt(math.pow(vs * math.cos(math.rad(angle)),2) + (2 * g * h))
-	
-	ab =  math.atan( math.sqrt(2 * g * h) / (vs * math.cos(math.rad(angle))) ) 
-	
-	l = ((math.pow(vs, 2) *  math.sin( math.rad(2 * angle))) + (math.pow(ve, 2) * math.sin( math.rad(2 * ab)))) / ( 2 * g)
-	print_message_to_user("ve="..ve)
-	print_message_to_user("ab="..ab)
-	print_message_to_user("l="..l)
-	print_message_to_user("h="..h)
-   end
-   -- if command == 2142 then
-		-- print_message_to_user("Horizontal="..value)
-		-- set_aircraft_draw_argument_value(19, value / -100)   
-		-- ROLL_F:set( value / -100)
-	-- end
-	
-	if command == 1507 then
-		if HMD_STATE:get() == 0 then
-			HMD_STATE:set(1)
-		else
-			HMD_STATE:set(0)
-		end
-   end
-   
-	if command == 2143 then
-		distance = 1 / math.sin( math.rad((value + 15)) )
-		print_message_to_user("distance="..(distance).."/anlge="..(value + 15))
-		-- set_aircraft_draw_argument_value(20, (value + 15) / 100)
-		YAM_F:set( distance)
-    end
-	
-	if command == device_commands.AltimeterSet then
-		alt_setting = alt_setting + value/10
-		if alt_setting > ALT_PRESSURE_MAX then
-			alt_setting = ALT_PRESSURE_MAX
-		elseif alt_setting < ALT_PRESSURE_MIN then
-			alt_setting = ALT_PRESSURE_MIN
-		end
-	end
-end
-
 
 function update_altimeter()
     local alt = sensor_data.getBarometricAltitude()*meters_to_feet
@@ -167,14 +114,7 @@ function update_altimeter()
     alt_10k:set((alt+alt_adj) % 100000)
     alt_1k:set((alt+alt_adj)  % 10000)
     alt_100s:set((alt+alt_adj) % 1000)
-	
-
 end
-
-
-
-
-
 
 function update_radar_altitude()
 	local pitch = (sensor_data.getPitch()*radian_to_degree)
@@ -188,6 +128,71 @@ function update_radar_altitude()
 	end
 end
 
+function update_lowalt()
+    local NowRalt = sensor_data.getRadarAltitude()
+
+    if NowRalt < 30 then
+        lowalt = lowalt + 0.1
+    end
+
+    if lowalt >= 1
+        then lowalt = 0
+    end
+
+    if NowRalt > 30 then
+        lowalt = 0
+    end
+end
+
+function toggleDebugInfo()
+    DEBUG_INFO_ENABLED:set(1 - DEBUG_INFO_ENABLED:get())
+    log.write("CH-53:Avionics:toggleDebugInfo()", log.INFO, "debug_info=%d", DEBUG_INFO_ENABLED:get())
+end
+
+function SetCommand(command, value)
+    -- print_message_to_user("command,value="..command) 
+    if command == 1501 then
+        h  = h0 + ( math.pow(vs * math.sin(math.rad(angle)), 2) / (2 * g))
+        ve = math.sqrt(math.pow(vs * math.cos(math.rad(angle)),2) + (2 * g * h))
+        
+        ab =  math.atan( math.sqrt(2 * g * h) / (vs * math.cos(math.rad(angle))) ) 
+        
+        l = ((math.pow(vs, 2) *  math.sin( math.rad(2 * angle))) + (math.pow(ve, 2) * math.sin( math.rad(2 * ab)))) / ( 2 * g)
+        print_message_to_user("ve="..ve)
+        print_message_to_user("ab="..ab)
+        print_message_to_user("l="..l)
+        print_message_to_user("h="..h)
+
+   -- elseif command == 2142 then
+		-- print_message_to_user("Horizontal="..value)
+		-- set_aircraft_draw_argument_value(19, value / -100)   
+		-- ROLL_F:set( value / -100)
+
+    elseif command == 1507 then
+		if HMD_STATE:get() == 0 then
+			HMD_STATE:set(1)
+		else
+			HMD_STATE:set(0)
+		end
+
+    elseif command == 2143 then
+		distance = 1 / math.sin( math.rad((value + 15)) )
+		print_message_to_user("distance="..(distance).."/anlge="..(value + 15))
+		-- set_aircraft_draw_argument_value(20, (value + 15) / 100)
+		YAM_F:set( distance)
+
+    elseif command == device_commands.AltimeterSet then
+		alt_setting = alt_setting + value/10
+		if alt_setting > ALT_PRESSURE_MAX then
+			alt_setting = ALT_PRESSURE_MAX
+		elseif alt_setting < ALT_PRESSURE_MIN then
+			alt_setting = ALT_PRESSURE_MIN
+		end
+
+    elseif command == EFM_commands.ToggleDebugInfo then
+        toggleDebugInfo()
+    end
+end
 
 function update()
 	update_lowalt()
@@ -211,34 +216,7 @@ function update()
 	set_aircraft_draw_argument_value(69,-1)-- to see if this affects ground crew
 	
 	LOWALT:set(lowalt)
-	
-	
 end
-
-
-function update_lowalt()
-
-	local NowRalt = sensor_data.getRadarAltitude()
-	
-if NowRalt < 30 then
-	lowalt = lowalt + 0.1
-end
-
-if lowalt >= 1
-	then lowalt = 0
-end
-
-if NowRalt > 30 then
-	lowalt = 0
-end
-
-end
-
- 
-
-	
- 
-
 
 
 need_to_be_closed = false
