+++ EFM/CH-53-EFM/CH53.h
@@ -8,8 +8,7 @@
 
 #include "CH53Electrics.h"
 #include "CH53Motion.h"
-//#include "CH53FuelSystem.h"
-#include "CH53Fuel.h"
+#include "CH53FuelSystem.h"
 #include "CH53AFCS.h"
 #include "CH53AFCSInput.h"
 #include "CH53AFCSAutopilot.h"
