+++ EFM/CH-53-EFM/CH53FuelSystem.h
@@ -1,114 +1,95 @@
-#ifndef _CH53_FUEL_SYSTEM_OBSOLETE_
-#define _CH53_FUEL_SYSTEM_OBSOLETE_
+#ifndef __CH53_FUEL_SYSTEM__
+#define __CH53_FUEL_SYSTEM__
 
 #include "CH53SimModule.h"
 
-// also calculation for the mass of the fuel and position of it
-
-// JP8 weight 6.65lbs/gal ?
-
 namespace CH53
 {
-
-    class FuelTank
-    {
-    public:
-
-        double volume; // capacity of tank
-        double fuel; // amount of fuel in tank
-
-        double x;
-        double y;
-        double z;
-
-        FuelTank(double _volume = 0, double _fuel = 0)
-            : volume(_volume)
-            , fuel(_fuel)
-            , x(0)
-            , y(0)
-            , z(0)
-        {}
-        ~FuelTank() {}
-
-        // add what is possible, return remaining if full
-        double addFuel(const double addition)
-        {
-            double space = volume - fuel;
-            if (space < addition)
-            {
-                fuel = volume; // set to max
-                return (addition - space); // overflow
-            }
-            fuel += addition;
-            return 0;
-        }
-
-        double decFuel(const double decrement)
-        {
-            if (fuel < decrement)
-            {
-                double tmp = decrement - fuel;
-                fuel = 0; // set to min
-                return tmp; // remaining
-            }
-            fuel -= decrement;
-            return 0;
-        }
-
-        double getSpace() const
-        {
-            return (volume - fuel);
-        }
-    };
-
     class FuelSystem : public SimModule
     {
-    protected:
-        bool is_unlimited_fuel;
-        double previous_usage;
+        enum FuelTank
+        {
+            FUEL_TANK_NO_1       = 0,
+            FUEL_TANK_NO_3       = 1,
+            FUEL_TANK_NO_2_LEFT  = 2,
+            FUEL_TANK_NO_2_RIGHT = 3,
+            FUEL_TANK_AUX_LEFT   = 4,
+            FUEL_TANK_AUX_RIGHT  = 5,
+        };
 
-        FuelTank MainTank;
-        FuelTank AuxTank;
+        struct Tank
+        {
+            double quantity_LBS;
+            double quantityPrevious_LBS;
+            double quantity_PCT;
+            double capacity_LBS;
+            Vec3   pos;
+        };
 
     public:
+        // consts
+        static constexpr double quantityTotalBINGO_LBS = 3313.0;
+        static constexpr double quantityTotalLOW_LBS   = 1165.0;
 
-        bool isIdleCutoff; // true means no fuel flow
+        // vars
+        FuelTank eSelectedTank                     = FuelTank::FUEL_TANK_NO_1;
+        double   capacityTotal_LBS                 = 0;
+        double   quantityTotal_LBS                 = 0;
+        double   quantityTotal_PCT                 = 0;
+        bool     bUnlimited                        = false;
+        bool     isIdleCutoff                      = false;  // true means no fuel flow
+        bool     isFuelFlow                        = false;
 
-        bool isFuelFlow;
+        std::vector<Tank>     fuelTanks            = {{0, 0, 0, 2040.0, Vec3(-1.0,  0.0, -2.3)}, //FUEL_TANK_NO_1,       300.0 gal, JP5
+                                                      {0, 0, 0, 2040.0, Vec3(-1.0,  0.0,  2.3)}, //FUEL_TANK_NO_3,       300.0 gal, JP5
+
+                                                      {0, 0, 0, 1281.8, Vec3( 1.0,  0.0, -2.3)}, //FUEL_TANK_NO_2_LEFT,  188.5 gal, JP5, cross-feed sections
+                                                      {0, 0, 0, 1281.8, Vec3( 1.0,  0.0,  2.3)}, //FUEL_TANK_NO_2_RIGHT, 188.5 gal, JP5, cross-feed sections
+
+                                                      {0, 0, 0, 3680.1, Vec3( 0.0,  0.0, -3.2)}, //FUEL_TANK_AUX_LEFT,   541.2 gal, JP5, (AUX)sponsons 
+                                                      {0, 0, 0, 3680.1, Vec3( 0.0,  0.0,  3.2)}, //FUEL_TANK_AUX_RIGHT,  541.2 gal, JP5, (AUX)sponsons 
+        };
 
         FuelSystem();
+
         virtual ~FuelSystem();
+
         virtual void vInit(bool hotStart, bool inAir);
+
         virtual void vRelease();
-
-        /*
-        // for weight-balance calculation,
-        // we need amount of fuel in each tank and position
-        double getFuelMass()
-    {
-            // JP8 weight 6.65lbs/gal ?
-            //return getInternalFuel() * weightconstant;
-            return 0;
-        }
-        */
-
-        // is low fuel indication
-        virtual bool isLowFuel() const;
-
-        // called on initialization
-        virtual void setInternalFuel(const double fuel);
-        virtual double getInternalFuel() const;
-        virtual void refuelAdd(const double fuel);
 
         virtual void setIdleCutOff(float value);
 
+        virtual double getTotalFuelCapacity();
 
-        virtual void setUnlimitedFuel(bool status);
+        virtual void setInternalFuelQuantity(double quantity_LBS);
 
-        virtual double getUsageSinceLastFrame() const;
-        virtual void clearUsageSinceLastFrame();
+        virtual void setExternalFuelQuantity(double quantity_LBS);
+
+        virtual double getInternalFuelQuantity();
+
+        virtual double getExternalFuelQuantity();
+
+        virtual double getTotalFuelQuantity();
+
+        virtual void addFuelQuantity(double quantity_LBS);
+
+        virtual double retrieveFuelFromTank(FuelSystem::FuelTank i, double requiredQuantityPartial_LBS);
+
+        virtual double retrieveFuel(double requiredQuantity_LBS);
+
+        virtual void setFuelUnlimited(bool flag);
+
+        virtual bool processMassChangeEvent(double& delta_mass,
+                                            double& delta_mass_pos_x,
+                                            double& delta_mass_pos_y,
+                                            double& delta_mass_pos_z,
+                                            double& delta_mass_moment_of_inertia_x,
+                                            double& delta_mass_moment_of_inertia_y,
+                                            double& delta_mass_moment_of_inertia_z);
+        virtual bool isLowFuel() const;
 
         virtual void vSimulate(struct Systems& systems, EDPARAM& cockpitAPI, double dt);
     };
 }
-#endif //_CH53_FUEL_SYSTEM_OBSOLETE_
+#endif //__EFM_FUEL_SYSTEM__
